add magit-display-buffer-fullframe-status-topleft-v1
This function is an alternative to
magit-display-buffer-fullframe-status-v1 that makes Magit buffers pop up
on the same side as they would with magit-display-buffer-traditional.

This is done by calling display-buffer-pop-up-window call and then
swapping the windows.  Another approach would be to make a variant of
display-buffer-pop-up-window that pops up on the desired side.  However,
doing this would require either relying on several internal window.el
functions or re-implementing these functions.  A third approach would be
to define a display-buffer-above-selected analog of
display-buffer-below-selected, but this has the same problem as the
second approach and also deviates from
magit-display-buffer-traditional's behavior in that it hardcodes a
horizontal split rather than respecting split-window-preferred-function.

Re: #2656